MyProject.Test/Application.Test/Services.Test/UserServiceTests.cs
using MyProject.Application.Services;
using MyProject.Application.Interfaces;
using MyProject.Application.DTO.User;
using Moq;
using System.Collections.Generic;
using System.Threading.Tasks;
using Xunit;

namespace MyProject.Tests.Services
{
    public class UserServiceTests
    {
        [Fact]
        public async Task GetAllAsync_ReturnsAllUsers()
        {
            // Arrange
            var mockUnitOfWork = new Mock<IUnitOfWork>();
            var userService = new UserService(mockUnitOfWork.Object);

            var expectedUsers = new List<UserResponseDTO>
            {
                /* Create mock user data or use actual data for testing */
            };
            mockUnitOfWork.Setup(u => u.Users.GetAllAsync()).ReturnsAsync(expectedUsers);

            // Act
            var result = await userService.GetAllAsync();

            // Assert
            Assert.NotNull(result);
            // Add more specific assertions to verify the returned data.
        }

        [Theory]
        [InlineData(1)] // Add more test cases if needed
        public async Task GetByIdAsync_ValidId_ReturnsUser(int userId)
        {
            // Arrange
            var mockUnitOfWork = new Mock<IUnitOfWork>();
            var userService = new UserService(mockUnitOfWork.Object);

            var expectedUser = new UserResponseDTO { /* Create a mock user or use actual data */ };
            mockUnitOfWork.Setup(u => u.Users.GetByIdAsync(userId)).ReturnsAsync(expectedUser);

            // Act
            var result = await userService.GetByIdAsync(userId);

            // Assert
            Assert.NotNull(result);
            // Add more specific assertions as needed
        }

        [Fact]
        public async Task CreateAsync_ValidUser_ReturnsUserId()
        {
            // Arrange
            var mockUnitOfWork = new Mock<IUnitOfWork>();
            var userService = new UserService(mockUnitOfWork.Object);

            var createUserRequest = new CreateUserRequestDTO { /* Create a mock user or use actual data */ };
            var expectedUserId = 1; // Replace with an expected user ID for testing purposes
            mockUnitOfWork.Setup(u => u.Users.CreateAsync(createUserRequest)).ReturnsAsync(expectedUserId);

            // Act
            var result = await userService.CreateAsync(createUserRequest);

            // Assert
            Assert.Equal(expectedUserId, result);
            // Add more specific assertions as needed
        }

        // Similarly, create test methods for UpdateAsync, DeleteAsync, and other scenarios

        // Example for UpdateAsync
        [Theory]
        [InlineData(1)] // Add more test cases if needed
        public async Task UpdateAsync_ValidId_ReturnsTrue(int userId)
        {
            // Arrange
            var mockUnitOfWork = new Mock<IUnitOfWork>();
            var userService = new UserService(mockUnitOfWork.Object);

            var updateUserRequest = new CreateUserRequestDTO { /* Create a mock user or use actual data */ };

            // Mock the UpdateAsync method to return true or false based on the test scenario
            mockUnitOfWork.Setup(u => u.Users.UpdateAsync(userId, updateUserRequest)).ReturnsAsync(true);

            // Act
            var result = await userService.UpdateAsync(userId, updateUserRequest);

            // Assert
            Assert.True(result);
            // Add more specific assertions as needed
        }

        // Example for DeleteAsync
        [Theory]
        [InlineData(1)] // Add more test cases if needed
        public async Task DeleteAsync_ValidId_ReturnsTrue(int userId)
        {
            // Arrange
            var mockUnitOfWork = new Mock<IUnitOfWork>();
            var userService = new UserService(mockUnitOfWork.Object);

            // Mock the DeleteAsync method to return true or false based on the test scenario
            mockUnitOfWork.Setup(u => u.Users.DeleteAsync(userId)).ReturnsAsync(true);

            // Act
            var result = await userService.DeleteAsync(userId);

            // Assert
            Assert.True(result);
            // Add more specific assertions as needed
        }
    }
}
